First draft of the logging system.
Result now has a log message queue. Normalizers add to the queue with Base::log_message which receives the log level (an enum) and the message, and automatically adds the Normalizer name. The maximum log level to output is set by Cycle, although it is currently hard coded to WARN). Output classes print the messages below the ignore level to their output stream.