A rudimentary plugin system.
Since Normalizer are now all shlibs, they can be dynamically loaded. This will also allow for multiple instances of the same Normalizer with different params (See #4). Params loading doesn't quite work yet because of the identification of plugins issues. Also, there needs to be a global option for the plugin base directory with a sensible default, maybe set from CMake.