Changes to Normalizer::name().
The name is no longer const char* but a string that can be set from outside. The normalizer must use this name as namespace for its parameters, so that different normalizers of the same type with different params can be used. By Normalizer::Base, it defaults to "Normalizer." If initialized from Applicator, it defaults to the library name (minus the lib- and -.so parts). Note that this invalidates the existing example cfg files for now, since those have the names in mixed case, whereas the libraries are all downcased names.